requirements/base.txt
aiohttp==3.9.5
aiosignal==1.3.1
asgiref==3.8.1
attrs==23.2.0
certifi==2024.2.2
charset-normalizer==3.3.2
Django==5.0.4
django-appconf==1.0.6
django-ckeditor-5==0.2.12
django-debug-toolbar==4.3.0
django-extensions==3.2.3
django-js-asset==2.2.0
django-mptt==0.16.0
django-paypal==2.0
frozenlist==1.4.1
geoip2==4.8.0
idna==3.7
maxminddb==2.6.1
multidict==6.0.5
mysqlclient==2.2.4
pillow==10.3.0
python-dotenv==1.0.1
python-slugify==8.0.4
pytz==2024.1
requests==2.31.0
six==1.16.0
sqlparse==0.5.0
stripe==9.3.0
text-unidecode==1.3
typing_extensions==4.11.0
tzdata==2024.1
urllib3==2.2.1
uuid==1.30
whitenoise==6.6.0
yarl==1.9.4
